[REFACTORING] Split IMAPServerOidcTest
